backend/requirements.txt
py-cpuinfo
gpustat
psutil
tornado
gunicorn
